Make MABM VPP token CRUD premium only
