+++ server/service/vpp.go
@@ -2,16 +2,12 @@ package service
 
 import (
 	"context"
-	"errors"
 	"io"
 	"mime/multipart"
 	"net/http"
 
 	"github.com/docker/go-units"
-	"github.com/fleetdm/fleet/v4/server/authz"
-	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
 	"github.com/fleetdm/fleet/v4/server/fleet"
-	"github.com/fleetdm/fleet/v4/server/mdm/apple/vpp"
 )
 
 //////////////////////////////////////////////////////////////////////////////
@@ -140,58 +136,12 @@ func uploadVPPTokenEndpoint(ctx context.Context, request interface{}, svc fleet.
 	return uploadVPPTokenResponse{Token: tok}, nil
 }
 
-func (svc *Service) UploadVPPToken(ctx context.Context, token io.ReadSeeker) (*fleet.VPPTokenDB, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.AppleCSR{}, fleet.ActionWrite); err != nil {
-		return nil, err
-	}
+func (svc *Service) UploadVPPToken(ctx context.Context, file io.ReadCloser) (*fleet.VPPTokenDB, error) {
+	// skipauth: No authorization check needed due to implementation returning
+	// only license error.
+	svc.authz.SkipAuthorization(ctx)
 
-	privateKey := svc.config.Server.PrivateKey
-	if testSetEmptyPrivateKey {
-		privateKey = ""
-	}
-
-	if len(privateKey) == 0 {
-		return nil, ctxerr.New(ctx, "Couldn't upload content token. Missing required private key. Learn how to configure the private key here: https://fleetdm.com/learn-more-about/fleet-server-private-key")
-	}
-
-	if token == nil {
-		return nil, ctxerr.Wrap(ctx, fleet.NewInvalidArgumentError("token", "Invalid token. Please provide a valid content token from Apple Business Manager."))
-	}
-
-	tokenBytes, err := io.ReadAll(token)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "reading VPP token")
-	}
-
-	locName, err := vpp.GetConfig(string(tokenBytes))
-	if err != nil {
-		var vppErr *vpp.ErrorResponse
-		if errors.As(err, &vppErr) {
-			// Per https://developer.apple.com/documentation/devicemanagement/app_and_book_management/app_and_book_management_legacy/interpreting_error_codes
-			if vppErr.ErrorNumber == 9622 {
-				return nil, ctxerr.Wrap(ctx, fleet.NewInvalidArgumentError("token", "Invalid token. Please provide a valid content token from Apple Business Manager."))
-			}
-		}
-		return nil, ctxerr.Wrap(ctx, err, "validating VPP token with Apple")
-	}
-
-	data := fleet.VPPTokenData{
-		Token:    string(tokenBytes),
-		Location: locName,
-	}
-
-	tok, err := svc.ds.InsertVPPToken(ctx, &data)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "writing VPP token to db")
-	}
-
-	if err := svc.NewActivity(ctx, authz.UserFromContext(ctx), fleet.ActivityEnabledVPP{
-		Location: locName,
-	}); err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "create activity for upload VPP token")
-	}
-
-	return tok, nil
+	return nil, fleet.ErrMissingLicense
 }
 
 ////////////////////////////////////////////////////
@@ -254,51 +204,11 @@ func patchVPPTokenRenewEndpoint(ctx context.Context, request interface{}, svc fl
 }
 
 func (svc *Service) UpdateVPPToken(ctx context.Context, tokenID uint, token io.ReadSeeker) (*fleet.VPPTokenDB, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.AppleCSR{}, fleet.ActionWrite); err != nil {
-		return nil, err
-	}
+	// skipauth: No authorization check needed due to implementation returning
+	// only license error.
+	svc.authz.SkipAuthorization(ctx)
 
-	privateKey := svc.config.Server.PrivateKey
-	if testSetEmptyPrivateKey {
-		privateKey = ""
-	}
-
-	if len(privateKey) == 0 {
-		return nil, ctxerr.New(ctx, "Couldn't upload content token. Missing required private key. Learn how to configure the private key here: https://fleetdm.com/learn-more-about/fleet-server-private-key")
-	}
-
-	if token == nil {
-		return nil, ctxerr.Wrap(ctx, fleet.NewInvalidArgumentError("token", "Invalid token. Please provide a valid content token from Apple Business Manager."))
-	}
-
-	tokenBytes, err := io.ReadAll(token)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "reading VPP token")
-	}
-
-	locName, err := vpp.GetConfig(string(tokenBytes))
-	if err != nil {
-		var vppErr *vpp.ErrorResponse
-		if errors.As(err, &vppErr) {
-			// Per https://developer.apple.com/documentation/devicemanagement/app_and_book_management/app_and_book_management_legacy/interpreting_error_codes
-			if vppErr.ErrorNumber == 9622 {
-				return nil, ctxerr.Wrap(ctx, fleet.NewInvalidArgumentError("token", "Invalid token. Please provide a valid content token from Apple Business Manager."))
-			}
-		}
-		return nil, ctxerr.Wrap(ctx, err, "validating VPP token with Apple")
-	}
-
-	data := fleet.VPPTokenData{
-		Token:    string(tokenBytes),
-		Location: locName,
-	}
-
-	tok, err := svc.ds.UpdateVPPToken(ctx, tokenID, &data)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "updating vpp token")
-	}
-
-	return tok, nil
+	return nil, fleet.ErrMissingLicense
 }
 
 ////////////////////////////////////////////////////
@@ -328,16 +238,11 @@ func patchVPPTokensTeams(ctx context.Context, request any, svc fleet.Service) (e
 }
 
 func (svc *Service) UpdateVPPTokenTeams(ctx context.Context, tokenID uint, teamIDs []uint) (*fleet.VPPTokenDB, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.AppleCSR{}, fleet.ActionWrite); err != nil {
-		return nil, err
-	}
+	// skipauth: No authorization check needed due to implementation returning
+	// only license error.
+	svc.authz.SkipAuthorization(ctx)
 
-	tok, err := svc.ds.UpdateVPPTokenTeams(ctx, tokenID, teamIDs)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "updating vpp token team")
-	}
-
-	return tok, nil
+	return nil, fleet.ErrMissingLicense
 }
 
 /////////////////////////////////////////
@@ -367,11 +272,11 @@ func getVPPTokens(ctx context.Context, request any, svc fleet.Service) (errorer,
 }
 
 func (svc *Service) GetVPPTokens(ctx context.Context) ([]*fleet.VPPTokenDB, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.AppleCSR{}, fleet.ActionRead); err != nil {
-		return nil, err
-	}
+	// skipauth: No authorization check needed due to implementation returning
+	// only license error.
+	svc.authz.SkipAuthorization(ctx)
 
-	return svc.ds.ListVPPTokens(ctx)
+	return nil, fleet.ErrMissingLicense
 }
 
 ///////////////////////////////////////////////
@@ -402,18 +307,9 @@ func deleteVPPToken(ctx context.Context, request any, svc fleet.Service) (errore
 }
 
 func (svc *Service) DeleteVPPToken(ctx context.Context, tokenID uint) error {
-	if err := svc.authz.Authorize(ctx, &fleet.AppleCSR{}, fleet.ActionWrite); err != nil {
-		return err
-	}
-	tok, err := svc.ds.GetVPPToken(ctx, tokenID)
-	if err != nil {
-		return ctxerr.Wrap(ctx, err, "getting vpp token")
-	}
-	if err := svc.NewActivity(ctx, authz.UserFromContext(ctx), fleet.ActivityDisabledVPP{
-		Location: tok.Location,
-	}); err != nil {
-		return ctxerr.Wrap(ctx, err, "create activity for delete VPP token")
-	}
+	// skipauth: No authorization check needed due to implementation returning
+	// only license error.
+	svc.authz.SkipAuthorization(ctx)
 
-	return svc.ds.DeleteVPPToken(ctx, tokenID)
+	return fleet.ErrMissingLicense
 }
